New modding patch : BaseFieldListUseFieldHost (#278)
The purpose of this patch if to allow `BaseField` and associated features (PAW controls, persistence, etc) to work when a custom `BaseField` is added to a `BaseFieldList` (ie, a `Part` or `PartModule`) with a `host` instance other than the `BaseFieldList` owner. This allow to dynamically add fields defined in another class to a `Part` or `PartModule` and to benefit from all the associated KSP sugar :
- PAW UI controls 
- Value and symmetry events
- Automatic persistence on the `Part` or `PartModule` hosting the `BaseFieldList`

Potential use cases for this are either :
- Having a part or module-level PAW item associated to and sharing the state of a common field, for example a field in a `KSPAddon`.
- Extending external (typically stock) modules with additional PAW UI controls and/or persisted fields.

Sample code demonstrating those use cases is provided in the source file, after the patch.