More balanced CommNetThrottling performance patch (#299)
Revamped mechanism that use an upper limit interval between updates in an in-game time set to 2.5 seconds, which is more acceptable for gameplay purposes. It also put an upper bound on the real time update frequency set to 50 hz (20 ms minimum interval), as there is no point in running at stupidly high update frequency on high frequency monitors or if the framerate isn't locked (manually or through V-Sync). That last setting can be increased to up to 500 or even 1000 ms in case the user doesn't care about simulation precision during timewarp and want to avoid the continuous performance hit in such cases. This PR also enable the patch by default.