perf: Optimize `MonoUtilities.RefreshContextWindows`
This calls `UnityEngine.Object.GetObjectsOfType` which is terribly slow.
It is not used anywhere in stock which is probably why it hasn't been
patched yet. However, several mods do make use of it.

On my KSP instance this saves > 1s on every scene switch across a few
different mods.

The improvement here is to maintain our own list of active
UIPartActionWindow instances and use that list instead of calling
`GetObjectsOfType`. It proactively removes all dead instances when one
is deleted. Any change that results in `UIPartActionWindow.OnDestroy`
not being called will result in a memory leak anyways so I suspect this
is a non-issue.