perf: Optimize away a `FindObjectOfType` call in `PQS.StartSphere` (#343)
* perf: Optimize away a FindObjectOfType call in PQS.StartSphere

`PQS.StartSphere` calls `FindObjectOfType(typeof(PQSCache))` intially.
However, PQSCache tracks the active instance in a static so we can just
use that instead.

This patch replaces the relevant call to FindObjectOfType with one that
just returns PQSCache.Instance if it is active. On my (admittedly rather
heavily modded) save this saves 1.2s off the scene switch time.

* Update CHANGELOG 

Added a new entry for unreleased improvements and a bug fix for FastAndFixedEnumExtensions.

---------

Co-authored-by: JonnyOThan <jonnyothan@gmail.com>