perf: Reparse the game database in parallel
During startup FastLoader reloads the whole game database. This seems to
take about 2s on my machine, though I have also see in vary quite a bit
in other people's profiles. This PR parallelizes the directory traversal
and config node parsing. The final result is that it now takes about
900ms to do.

Tuning the parallelism for stock has actually been rather difficult.
What I have here is that it recurses 3 directories deep, so:
`GameData\ModName\InnerFolder`. This is probably undertuned for stock,
since most of the work is in `GameData\Squad\Parts` but going deeper
would likely result in too many tasks flying around (and, tbh, I expect
that to already be the case on a large modded install already).